+++ libft/ft_is_integer.c
@@ -6,13 +6,13 @@
 /*   By: akusniak <akusniak@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/13 12:24:01 by akusniak          #+#    #+#             */
-/*   Updated: 2023/02/13 13:51:27 by akusniak         ###   ########.fr       */
+/*   Updated: 2023/03/05 15:31:34 by akusniak         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_is_integer(long number)
+int	ft_is_integer(long long number)
 {
 	if (number < -2147483648 || number > 2147483647)
 		return (0);
